+++ agnocomplete/core.py
@@ -18,6 +18,8 @@ from .constants import AGNOCOMPLETE_MAX_PAGESIZE
 from .constants import AGNOCOMPLETE_DEFAULT_QUERYSIZE
 from .constants import AGNOCOMPLETE_MIN_QUERYSIZE
 from .exceptions import AuthenticationRequiredAgnocompleteException
+from .exceptions import SkipItem
+from .exceptions import ItemNotFound
 
 
 logger = logging.getLogger(__name__)
@@ -618,7 +620,10 @@ class AgnocompleteUrlProxy(with_metaclass(ABCMeta, AgnocompleteBase)):
         result = []
         for item in http_result:
             # Eventual result reshaping.
-            result.append(self.item(item))
+            try:
+                result.append(self.item(item))
+            except SkipItem:
+                continue
         return result
 
     def selected(self, ids):
@@ -638,3 +643,27 @@ class AgnocompleteUrlProxy(with_metaclass(ABCMeta, AgnocompleteBase)):
                             )
                         )
         return data
+
+    def validate(self, value):
+        """
+        From a value available on the remote server, the method returns the
+        complete item matching the value.
+        If case the value is not available on the server side or filtered
+        through :meth:`item`, the class:`agnocomplete.exceptions.ItemNotFound`
+        is raised.
+        """
+
+        url = self.get_item_url(value)
+        try:
+            data = self.http_call(url=url)
+        except requests.HTTPError:
+            raise ItemNotFound()
+
+        data = self.get_http_result(data)
+
+        try:
+            self.item(data)
+        except SkipItem:
+            raise ItemNotFound()
+
+        return value