+++ agnocomplete/exceptions.py
@@ -30,3 +30,17 @@ class HTTPError(HTTPError):
     Occurs when the 3rd party API returns an error code
     """
     pass
+
+
+class SkipItem(Exception):
+    """
+    Occurs when Item has to be skipped when building the final Payload
+    """
+    pass
+
+
+class ItemNotFound(Exception):
+    """
+    Occurs while searching an unexisting item on autocomplete choices.
+    """
+    pass